+++ CHANGELOG.md
@@ -8,12 +8,14 @@ Bug fix:
 - [#747](https://github.com/elysiajs/elysia/issues/747) mapResponse inside mapError override error value
 - [#812](https://github.com/elysiajs/elysia/issues/812) check for minItems length before array validation
 - [#833](https://github.com/elysiajs/elysia/issues/833) cookie signing doesn't work in dynamic mode
+- [#859](https://github.com/elysiajs/elysia/issues/859) clean non-root additionalProperties
 - [#924](https://github.com/elysiajs/elysia/issues/924) decode path param
 - [#1274](https://github.com/elysiajs/elysia/issues/1274) support .use(undefined | false) for conditional plugin
 - [#1276](https://github.com/elysiajs/elysia/issues/1276) mapResponse with set inference produce invalid instruction
 - [#1268](https://github.com/elysiajs/elysia/issues/1268) using number instead of stringifed value for reporting validation error
 - [#1323](https://github.com/elysiajs/elysia/issues/1323) don't duplicate error from plugin
 - dynamic handle should handle named parser
+- hand
 
 Change:
 - afterResponse now call after response by scheduling setImmediate